rename: Modify gallery component name as photo
