--- apps/web/src/features/gallery/feed/index.ts
@@ -1,1 +1,0 @@
-export { default as FeedList } from './ui/FeedList';
