+++ apps/web/src/features/gallery/photos/index.ts
@@ -1,0 +1,1 @@
+export { default as PhotoGrid } from './ui/PhotoGrid';
