+++ apps/web/src/views/gallery/ui/GalleryPage.tsx
@@ -1,4 +1,4 @@
-import { FeedList } from '@/fsd/features/gallery/feed';
+import { PhotoGrid } from '@/fsd/features/gallery/photos';
 
 import { Container } from '@jung/design-system/components';
 // import { Suspense } from 'react';
@@ -7,7 +7,7 @@ const GalleryPage = () => {
 	return (
 		<Container>
 			{/* <Suspense fallback={<GalleryCardSkeleton count={12} />}> */}
-			<FeedList />
+			<PhotoGrid />
 			{/* </Suspense> */}
 		</Container>
 	);
