+++ packages/shared/types/index.ts
@@ -2,6 +2,6 @@ export * from './comment';
 export * from './toast';
 export type { Post } from './post';
 export { PostSchema } from './post';
-export type { Feed } from './feed';
-export { FeedSchema } from './feed';
-export type { CustomFeed as CustomPhoto } from './feed';
+export type { Photo } from './photo';
+export { PhotoSchema } from './photo';
+export type { CustomPhoto } from './photo';
